+++ site/app/Application/LocaleLinkGenerator.php
@@ -4,6 +4,7 @@ declare(strict_types = 1);
 namespace MichalSpacekCz\Application;
 
 use Contributte\Translation\Translator;
+use MichalSpacekCz\Application\Routing\RouterFactory;
 use MichalSpacekCz\ShouldNotHappenException;
 use Nette\Application\IPresenterFactory;
 use Nette\Application\LinkGenerator;
