+++ site/app/Application/Routing/RouterFactoryRoute.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace MichalSpacekCz\Application;
+namespace MichalSpacekCz\Application\Routing;
 
 class RouterFactoryRoute
 {
