+++ site/psalm.xml
@@ -45,7 +45,7 @@
 				<referencedMethod name="/.*Presenter::inject.*/" />
 				<referencedMethod name="/.*Test::get.*/" /> <!-- Methods used in @dataProvider -->
 				<referencedMethod name="/.*Test::test.*/" />
-				<referencedMethod name="MichalSpacekCz\Application\RouterFactory::createRouter" /> <!-- Used in services.neon -->
+				<referencedMethod name="MichalSpacekCz\Application\Routing\RouterFactory::createRouter" /> <!-- Used in services.neon -->
 				<referencedMethod name="MichalSpacekCz\DateTime\DateTimeFormatter::localeMonth" /> <!-- Used in templates -->
 				<referencedMethod name="/^MichalSpacekCz\\Templating\\Filters::(staticUrl|staticImageUrl)$/" /> <!-- Used in templates -->
 				<referencedMethod name="/^MichalSpacekCz\\Test\\Http\\Request::(set|add).*/" /> <!-- Not used but keep them just in case -->
